Algo improved: map of sorted boxes that will be updated. GetBoxesAndX() now faster.

+++ src/main.go
@@ -74,7 +74,7 @@ func main() {
 			for {
 				fmt.Scanf("%d", &input)
 				if input >= 0 && input <= 3 {
-					sokoban.Move(input)
+					sokoban.Move(int8(input))
 					sokoban.Print()
 				} else {
 					sokoban.UndoStep()
